+++ example/lib/main.dart
@@ -1,9 +1,0 @@
-import 'package:wordpress_api/wordpress_api.dart';
-
-void main() async {
-  final wp = WordPressAPI('mwebantu.com');
-  final posts = (await wp.search("lungu")).data;
-  for (final post in posts) {
-    print(post);
-  }
-}