+++ lib/constants.dart
@@ -1,2 +1,0 @@
-final WP_NAMESPACE = "wp/v2/";
-final WOO_NAMESPACE = "wc/v3/";