+++ lib/src/constants.dart
@@ -1,0 +1,2 @@
+const wpNamespace = "wp/v2/";
+const wooNamespace = "wc/v3/";