+++ lib/src/enums.dart
@@ -1,0 +1,34 @@
+enum WPEndpoint {
+  /// /wp/v2/categories
+  categories,
+
+  /// /wp/v2/comments
+  comments,
+
+  /// /custom-endpoint
+  custom,
+
+  /// /wp/v2/media
+  media,
+
+  /// /wp/v2/pages
+  pages,
+
+  /// /wp/v2/posts
+  posts,
+
+  /// /wp/v2/search
+  search,
+
+  /// /wp/v2/settings
+  settings,
+
+  /// /wp/v2/tags
+  tags,
+
+  /// /wp/v2/taxonomies
+  taxonomies,
+
+  /// /wp/v2/users
+  users
+}