+++ lib/src/schemas.dart
@@ -1,0 +1,12 @@
+library schemas;
+
+export 'schemas/category.dart';
+export 'schemas/comment.dart';
+export 'schemas/media.dart';
+export 'schemas/page.dart';
+export 'schemas/post.dart';
+export 'schemas/search.dart';
+export 'schemas/settings.dart';
+export 'schemas/tag.dart';
+export 'schemas/taxonomy.dart';
+export 'schemas/user.dart';