+++ lib/src/schemas/media.dart
@@ -1,15 +1,16 @@
 import 'package:json_annotation/json_annotation.dart';
+import 'package:wordpress_api/src/schemas/media_details.dart';
 
-part 'page.g.dart';
+part 'media.g.dart';
 
-/// WordPress Page Endpoint Model /wp/v2/pages
+/// WordPress Media Endpoint Model /wp/v2/media
 @JsonSerializable()
-class PageSchema {
+class MediaSchema {
   ///The date the object was published, in the site's timezone.
-  final List date;
+  final String date;
 
   ///The date the object was published, as GMT.
-  final List dateGmt;
+  final String dateGmt;
 
   ///The globally unique identifier for the object.
   final Map<String, dynamic> guid;
@@ -24,7 +25,7 @@ class PageSchema {
   final String modified;
 
   ///The date the object was last modified, as GMT.
-  final String modifitedGmt;
+  final String modifiedGmt;
 
   ///An alphanumeric identifier for the object unique to its type.
   final String slug;
@@ -37,32 +38,17 @@ class PageSchema {
   ///Type of Post for the object.
   final String type;
 
-  ///A password to protect access to the content and excerpt.
-  final String password;
-
   ///Permalink template for the object.
   final String permalinkTemplate;
 
   ///Slug automatically generated from the object title.
   final String generatedSlug;
 
-  ///The ID for the parent of the object.
-  final int parent;
-
   ///The title for the object.
   final Map<String, dynamic> title;
 
-  ///The content for the object.
-  final Map<String, dynamic> content;
-
   ///The ID for the author of the object.
   final int author;
-
-  ///The excerpt for the object.
-  final Map<String, dynamic> excerpt;
-
-  ///The ID of the featured media for the object.
-  final int featuredMedia;
 
   ///Whether or not comments are open on the object.
   ///
@@ -74,46 +60,80 @@ class PageSchema {
   /// One of: "open", "closed"
   final String pingStatus;
 
-  ///The order of the object in relation to other object of its type.
-  final int menuOrder;
-
   ///Meta fields.
-  final Map<String, dynamic> meta;
+  final List meta;
 
   ///The theme file to use to display the object.
   final String template;
 
-  PageSchema({
+  ///Alternative text to display when attachment is not displayed.
+  final String altText;
+
+  ///The attachment caption.
+  final Map<String, dynamic> caption;
+
+  ///The attachment description.
+  final Map<String, dynamic> description;
+
+  ///Attachment type.
+  ///
+  /// One of: "image", "file"
+  final String mediaType;
+
+  ///The attachment MIME type.
+  final String mimeType;
+
+  ///Details about the media file, specific to its type.
+  final MediaDetailsSchema mediaDetails;
+
+  ///The ID for the associated post of the attachment.
+  final int post;
+
+  ///URL to the original attachment file.
+  final String sourceUrl;
+
+  ///List of the missing image sizes of the attachment.
+  final List missingImageSizes;
+
+  MediaSchema({
+    this.permalinkTemplate,
+    this.generatedSlug,
+    this.missingImageSizes,
+    this.id,
     this.date,
     this.dateGmt,
     this.guid,
-    this.id,
-    this.link,
     this.modified,
-    this.modifitedGmt,
+    this.modifiedGmt,
     this.slug,
     this.status,
     this.type,
-    this.password,
-    this.permalinkTemplate,
-    this.generatedSlug,
-    this.parent,
+    this.link,
     this.title,
-    this.content,
     this.author,
-    this.excerpt,
-    this.featuredMedia,
     this.commentStatus,
     this.pingStatus,
-    this.menuOrder,
-    this.meta,
     this.template,
+    this.meta,
+    this.description,
+    this.caption,
+    this.altText,
+    this.mediaType,
+    this.mimeType,
+    this.mediaDetails,
+    this.post,
+    this.sourceUrl,
   });
 
-  /// Generate PageSchema Model from JSON
-  factory PageSchema.fromJson(Map<String, dynamic> data) =>
-      _$PageSchemaFromJson(data);
+  @override
+  String toString() {
+    return sourceUrl;
+  }
 
-  /// Convert PageSchema Model to JSON
-  Map<String, dynamic> toJson() => _$PageSchemaToJson(this);
+  /// Generate MediaSchema Model from JSON
+  factory MediaSchema.fromJson(Map<String, dynamic> data) =>
+      _$MediaSchemaFromJson(data);
+
+  /// Convert MediaSchema Model to JSON
+  Map<String, dynamic> toJson() => _$MediaSchemaToJson(this);
 }