+++ lib/src/schemas/media.g.dart
@@ -1,43 +1,47 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'post.dart';
+part of 'media.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-PostSchema _$PostSchemaFromJson(Map<String, dynamic> json) {
-  return PostSchema(
+MediaSchema _$MediaSchemaFromJson(Map<String, dynamic> json) {
+  return MediaSchema(
+    permalinkTemplate: json['permalink_template'] as String,
+    generatedSlug: json['generated_slug'] as String,
+    missingImageSizes: json['missing_image_sizes'] as List,
+    id: json['id'] as int,
     date: json['date'] as String,
     dateGmt: json['date_gmt'] as String,
     guid: json['guid'] as Map<String, dynamic>,
-    id: json['id'] as int,
-    link: json['link'] as String,
     modified: json['modified'] as String,
     modifiedGmt: json['modified_gmt'] as String,
     slug: json['slug'] as String,
     status: json['status'] as String,
     type: json['type'] as String,
-    password: json['password'] as String,
-    permalinkTemplate: json['permalink_template'] as String,
-    generatedSlug: json['generated_slug'] as String,
+    link: json['link'] as String,
     title: json['title'] as Map<String, dynamic>,
-    content: json['content'] as Map<String, dynamic>,
     author: json['author'] as int,
-    excerpt: json['excerpt'] as Map<String, dynamic>,
-    featuredMedia: json['featured_media'] as int,
     commentStatus: json['comment_status'] as String,
     pingStatus: json['ping_status'] as String,
-    format: json['format'] as String,
-    meta: json['meta'] as Map<String, dynamic>,
-    sticky: json['sticky'] as bool,
     template: json['template'] as String,
-    categories: json['categories'] as List,
-    tags: json['tags'] as List,
+    meta: json['meta'] as List,
+    description: json['description'] as Map<String, dynamic>,
+    caption: json['caption'] as Map<String, dynamic>,
+    altText: json['alt_text'] as String,
+    mediaType: json['media_type'] as String,
+    mimeType: json['mime_type'] as String,
+    mediaDetails: json['media_details'] == null
+        ? null
+        : MediaDetailsSchema.fromJson(
+            json['media_details'] as Map<String, dynamic>),
+    post: json['post'] as int,
+    sourceUrl: json['source_url'] as String,
   );
 }
 
-Map<String, dynamic> _$PostSchemaToJson(PostSchema instance) =>
+Map<String, dynamic> _$MediaSchemaToJson(MediaSchema instance) =>
     <String, dynamic>{
       'date': instance.date,
       'date_gmt': instance.dateGmt,
@@ -49,20 +53,21 @@ Map<String, dynamic> _$PostSchemaToJson(PostSchema instance) =>
       'slug': instance.slug,
       'status': instance.status,
       'type': instance.type,
-      'password': instance.password,
       'permalink_template': instance.permalinkTemplate,
       'generated_slug': instance.generatedSlug,
       'title': instance.title,
-      'content': instance.content,
       'author': instance.author,
-      'excerpt': instance.excerpt,
-      'featured_media': instance.featuredMedia,
       'comment_status': instance.commentStatus,
       'ping_status': instance.pingStatus,
-      'format': instance.format,
       'meta': instance.meta,
-      'sticky': instance.sticky,
       'template': instance.template,
-      'categories': instance.categories,
-      'tags': instance.tags,
+      'alt_text': instance.altText,
+      'caption': instance.caption,
+      'description': instance.description,
+      'media_type': instance.mediaType,
+      'mime_type': instance.mimeType,
+      'media_details': instance.mediaDetails,
+      'post': instance.post,
+      'source_url': instance.sourceUrl,
+      'missing_image_sizes': instance.missingImageSizes,
     };