+++ lib/src/schemas/post.g.dart
@@ -1,44 +1,46 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'media.dart';
+part of 'post.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-MediaSchema _$MediaSchemaFromJson(Map<String, dynamic> json) {
-  return MediaSchema(
-    permalinkTemplate: json['permalink_template'] as String,
-    generatedSlug: json['generated_slug'] as String,
-    missingImageSizes: json['missing_image_sizes'] as List,
-    id: json['id'] as int,
+PostSchema _$PostSchemaFromJson(Map<String, dynamic> json) {
+  return PostSchema(
     date: json['date'] as String,
     dateGmt: json['date_gmt'] as String,
     guid: json['guid'] as Map<String, dynamic>,
+    id: json['id'] as int,
+    link: json['link'] as String,
     modified: json['modified'] as String,
     modifiedGmt: json['modified_gmt'] as String,
     slug: json['slug'] as String,
     status: json['status'] as String,
     type: json['type'] as String,
-    link: json['link'] as String,
+    password: json['password'] as String,
+    permalinkTemplate: json['permalink_template'] as String,
+    generatedSlug: json['generated_slug'] as String,
     title: json['title'] as Map<String, dynamic>,
+    content: json['content'] as Map<String, dynamic>,
     author: json['author'] as int,
+    excerpt: json['excerpt'] as Map<String, dynamic>,
+    featuredMedia: json['featured_media'] as int,
     commentStatus: json['comment_status'] as String,
     pingStatus: json['ping_status'] as String,
-    template: json['template'] as String,
+    format: json['format'] as String,
     meta: json['meta'] as Map<String, dynamic>,
-    description: json['description'] as Map<String, dynamic>,
-    caption: json['caption'] as Map<String, dynamic>,
-    altText: json['alt_text'] as String,
-    mediaType: json['media_type'] as String,
-    mimeType: json['mime_type'] as String,
-    mediaDetails: json['media_details'] as Map<String, dynamic>,
-    post: json['post'] as int,
-    sourceUrl: json['source_url'] as String,
+    sticky: json['sticky'] as bool,
+    template: json['template'] as String,
+    categories: json['categories'] as List,
+    tags: json['tags'] as List,
+    embedded: json['_embedded'] == null
+        ? null
+        : EmbeddedSchema.fromJson(json['_embedded'] as Map<String, dynamic>),
   );
 }
 
-Map<String, dynamic> _$MediaSchemaToJson(MediaSchema instance) =>
+Map<String, dynamic> _$PostSchemaToJson(PostSchema instance) =>
     <String, dynamic>{
       'date': instance.date,
       'date_gmt': instance.dateGmt,
@@ -50,21 +52,21 @@ Map<String, dynamic> _$MediaSchemaToJson(MediaSchema instance) =>
       'slug': instance.slug,
       'status': instance.status,
       'type': instance.type,
+      'password': instance.password,
       'permalink_template': instance.permalinkTemplate,
       'generated_slug': instance.generatedSlug,
       'title': instance.title,
+      'content': instance.content,
       'author': instance.author,
+      'excerpt': instance.excerpt,
+      'featured_media': instance.featuredMedia,
       'comment_status': instance.commentStatus,
       'ping_status': instance.pingStatus,
+      'format': instance.format,
       'meta': instance.meta,
+      'sticky': instance.sticky,
       'template': instance.template,
-      'alt_text': instance.altText,
-      'caption': instance.caption,
-      'description': instance.description,
-      'media_type': instance.mediaType,
-      'mime_type': instance.mimeType,
-      'media_details': instance.mediaDetails,
-      'post': instance.post,
-      'source_url': instance.sourceUrl,
-      'missing_image_sizes': instance.missingImageSizes,
+      'categories': instance.categories,
+      'tags': instance.tags,
+      '_embedded': instance.embedded,
     };