+++ lib/src/utils.dart
@@ -1,0 +1,4 @@
+
+import 'package:logger/logger.dart';
+
+Logger get logger => Logger();