+++ lib/utils.dart
@@ -1,5 +1,0 @@
-import 'package:logger/logger.dart';
-
-abstract class Utils {
-  static Logger get logger => Logger();
-}